restructure model building script
